[SECURITY-3485]
